+++ Modules/Microsoft365DSC/DSCResources/MSFT_O365SearchAndIntelligenceConfigurations/MSFT_O365SearchAndIntelligenceConfigurations.psm1
@@ -34,9 +34,32 @@ function Get-TargetResource
         $Credential,
 
         [Parameter()]
+        [System.String]
+        $ApplicationId,
+
+        [Parameter()]
+        [System.String]
+        $TenantId,
+
+        [Parameter()]
+        [System.String]
+        $CertificateThumbprint,
+
+        [Parameter()]
+        [System.String]
+        $CertificatePath,
+
+        [Parameter()]
+        [System.Management.Automation.PSCredential]
+        $CertificatePassword,
+
+        [Parameter()]
         [System.String[]]
         $AccessTokens
     )
+
+    $ConnectionMode = New-M365DSCConnection -Workload 'ExchangeOnline' `
+        -InboundParameters $PSBoundParameters
 
     $ConnectionMode = New-M365DSCConnection -Workload 'MicrosoftGraph' `
         -InboundParameters $PSBoundParameters
@@ -91,14 +114,21 @@ function Get-TargetResource
             }
         }
 
+        $MeetingInsightsResponse = Get-MeetingInsightsSettings
+        $MeetingInsightsValue    = [Boolean]::Parse($MeetingInsightsResponse.Split(':')[1].Trim())
+
         return @{
-            IsSingleInstance                      = 'Yes'
-            ItemInsightsIsEnabledInOrganization   = $ItemInsights.IsEnabledInOrganization
-            ItemInsightsDisabledForGroup          = $itemInsightsDisabledForGroupValue
-            PersonInsightsIsEnabledInOrganization = $PersonInsights.IsEnabledInOrganization
-            PersonInsightsDisabledForGroup        = $PersonInsightsDisabledForGroupValue
-            Credential                            = $Credential
-            AccessTokens                          = $AccessTokens
+            IsSingleInstance                       = 'Yes'
+            ItemInsightsIsEnabledInOrganization    = $ItemInsights.IsEnabledInOrganization
+            ItemInsightsDisabledForGroup           = $itemInsightsDisabledForGroupValue
+            MeetingInsightsIsEnabledInOrganization = $MeetingInsightsValue
+            PersonInsightsIsEnabledInOrganization  = $PersonInsights.IsEnabledInOrganization
+            PersonInsightsDisabledForGroup         = $PersonInsightsDisabledForGroupValue
+            Credential                             = $Credential
+            ApplicationId                          = $ApplicationId
+            TenantId                               = $TenantId
+            CertificateThumbprint                  = $CertificateThumbprint
+            AccessTokens                           = $AccessTokens
         }
     }
     catch
@@ -146,6 +176,26 @@ function Set-TargetResource
         [Parameter()]
         [System.Management.Automation.PSCredential]
         $Credential,
+
+        [Parameter()]
+        [System.String]
+        $ApplicationId,
+
+        [Parameter()]
+        [System.String]
+        $TenantId,
+
+        [Parameter()]
+        [System.String]
+        $CertificateThumbprint,
+
+        [Parameter()]
+        [System.String]
+        $CertificatePath,
+
+        [Parameter()]
+        [System.Management.Automation.PSCredential]
+        $CertificatePassword,
 
         [Parameter()]
         [System.String[]]
@@ -227,6 +277,11 @@ function Set-TargetResource
     Write-Verbose -Message 'Updating settings for Person Insights'
     Update-MgBetaOrganizationSettingPersonInsight @PersonInsightsUpdateParams | Out-Null
     #endregion
+
+    if ($null -ne $MeetingInsightsIsEnabledInOrganization)
+    {
+        Set-MeetingInsightsSettings -Enabled $MeetingInsightsIsEnabledInOrganization | Out-Null
+    }
 }
 
 function Test-TargetResource
@@ -263,6 +318,26 @@ function Test-TargetResource
         [Parameter()]
         [System.Management.Automation.PSCredential]
         $Credential,
+
+        [Parameter()]
+        [System.String]
+        $ApplicationId,
+
+        [Parameter()]
+        [System.String]
+        $TenantId,
+
+        [Parameter()]
+        [System.String]
+        $CertificateThumbprint,
+
+        [Parameter()]
+        [System.String]
+        $CertificatePath,
+
+        [Parameter()]
+        [System.Management.Automation.PSCredential]
+        $CertificatePassword,
 
         [Parameter()]
         [System.String[]]
@@ -309,6 +384,26 @@ function Export-TargetResource
         $Credential,
 
         [Parameter()]
+        [System.String]
+        $ApplicationId,
+
+        [Parameter()]
+        [System.String]
+        $TenantId,
+
+        [Parameter()]
+        [System.String]
+        $CertificateThumbprint,
+
+        [Parameter()]
+        [System.String]
+        $CertificatePath,
+
+        [Parameter()]
+        [System.Management.Automation.PSCredential]
+        $CertificatePassword,
+
+        [Parameter()]
         [System.String[]]
         $AccessTokens
     )
@@ -335,9 +430,12 @@ function Export-TargetResource
         }
 
         $Params = @{
-            IsSingleInstance = 'Yes'
-            Credential       = $Credential
-            AccessTokens     = $AccessTokens
+            IsSingleInstance      = 'Yes'
+            Credential            = $Credential
+            AccessTokens          = $AccessTokens
+            ApplicationId         = $ApplicationId
+            TenantId              = $TenantId
+            CertificateThumbprint = $CertificateThumbprint
         }
 
         $Results = Get-TargetResource @Params